Safari-611.1.10/LayoutTests/accessibility/mac/option-with-arialabel-expected.txt


This tests that aria-label can be set on options.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS popup.stringValue is 'AXValue: Fahrenheit'
PASS selectbox.childAtIndex(0).stringValue is 'AXValue: Fahrenheit'
PASS selectbox.childAtIndex(1).stringValue is 'AXValue: Celsius'
PASS successfullyParsed is true

TEST COMPLETE
